Removed inline from ctype functions
We'll let the compiler deal with it if it wants to. The rationale is that they're not usually called functions, and even if they are, they're not exactly mission critical. Also, every inline option explored posed many issues, some breaking compilation, some just including every function in every file, which is unacceptable. In the end, we just have to give up, at least for now.
As a bonus, we added toupper and tolower